+++ base/exports.jl
@@ -855,7 +855,6 @@ export
     readline,
     readlines,
     readuntil,
-    readuntil!,
     redirect_stdio,
     redirect_stderr,
     redirect_stdin,